Revert "feat: custom user agent Auth changes for UI handoff (#11606)"
This reverts commit 2790f0421b70977547378108e890d442637a45ba.
