+++ packages/auth/__tests__/auth-attribute-test.ts
@@ -10,7 +10,6 @@ import {
 } from 'amazon-cognito-identity-js';
 
 import { AuthOptions } from '../src/types';
-import { InternalAuthClass } from '../src/internals/InternalAuth';
 
 const authOptions: AuthOptions = {
 	userPoolId: 'us-west-2_0xxxxxxxx',
@@ -23,7 +22,7 @@ const authOptions: AuthOptions = {
 describe('User-Attribute-validation', () => {
 	it('Check-non-verified-attributes', async () => {
 		const spyonAuthUserAttributes = jest
-			.spyOn(InternalAuthClass.prototype, 'userAttributes')
+			.spyOn(Auth.prototype, 'userAttributes')
 			.mockImplementation((user: CognitoUser) => {
 				const emailAttribute = new CognitoUserAttribute({
 					Name: 'email',
