+++ packages/auth/package.json
@@ -43,8 +43,7 @@
   "files": [
     "lib",
     "lib-esm",
-    "src",
-    "internals"
+    "src"
   ],
   "dependencies": {
     "@aws-amplify/core": "5.8.3",
@@ -61,7 +60,7 @@
       "name": "Auth (top-level class)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, Auth }",
-      "limit": "56.29 kB"
+      "limit": "55.73 kB"
     }
   ],
   "jest": {
