+++ packages/auth/tslint.json
@@ -39,12 +39,7 @@
 			"allow-snake-case",
 			"allow-leading-underscore"
 		],
-		"semicolon": [
-			true,
-			"always",
-			"ignore-interfaces",
-			"ignore-bound-class-methods"
-		]
+		"semicolon": [true, "always", "ignore-interfaces"]
 	},
 	"rulesDirectory": []
 }
