+++ packages/datastore/package.json
@@ -73,7 +73,7 @@
 			"name": "DataStore (top-level class)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, DataStore }",
-			"limit": "138.27 kB"
+			"limit": "137.68 kB"
 		}
 	],
 	"jest": {
